Renaming DeleteSiteDialog to DeleteSiteDialogFragment
